add ChatApi
